k8s-agent/requirements.txt
# 核心依赖
python-dotenv>=0.19.0
openai>=1.0.0
kubernetes>=26.1.0
click>=8.0.0
rich>=10.0.0

# 开发依赖
pytest>=7.0.0
black>=22.0.0
isort>=5.0.0
flake8>=4.0.0
mypy>=0.900 